requirements.txt
django===>1.11
dj-database-url
# social-auth-app-django
psycopg2

Pillow
pyyaml

django-model-utils
graphene-django>=1.0
djangorestframework
django-filter

django-spaghetti-and-meatballs

waitress
whitenoise
django-cors-middleware


python-decouple